requirements.txt
openai
python-dotenv
spacy
vaderSentiment
textblob
pymongo
